Implement prediction interface
