+++ backend/oracle.py
@@ -2,49 +2,10 @@ import boto3
 import json
 import os
 import time
+
+import status
 from db import db
-
-status_to_orders = {
-    -1: 0,
-    0: 1,
-    1: -1,
-}
-
-
-def _build_status(output):
-    return [
-        {
-            "VaTech": {
-                "soc": 10.0,
-                "billing_demand": -1.0,
-                "regd_ts": 1430496000.0,
-                "battery": 0.0,
-                "emergency_time": 1.333333,
-                "ts": time.time(),
-                "other": {
-                    "plug1": {
-                        "Power Sum": None
-                    },
-                    "compressor": {
-                        "Power Sum": None
-                    },
-                    "airhandler": {
-                        "Power Sum": None
-                    },
-                    "plug2": {
-                        "Power Sum": None
-                    },
-                    "lighting": {
-                        "Power Sum": None
-                    }
-                },
-                "usage": 0.0,
-                "output": output,
-                "target_output": 1.0,
-                "pv_production": 0.0
-            }
-        }
-    ]
+from predict import predict
 
 
 def handler(datapoints, context):
@@ -55,13 +16,15 @@ def handler(datapoints, context):
         region = 'us-west-2'
         dynamodb_url = None
 
+    prediction_strategy = predict.get_strategy()
+
     orders = None
 
     for datapoint in datapoints:
         device_id = datapoint.keys()[0]
         data = datapoint[device_id]
         _write(device_id, data, table_name, region, dynamodb_url)
-        orders = get_orders(data)
+        orders = prediction_strategy.predict(device_id, data)
 
     return orders
 
@@ -78,11 +41,3 @@ def _write(device_id, data, table_name, region, dynamodb_url):
     }
 
     dynamodb_client.put_item(TableName=table_name, Item=item)
-
-
-def get_orders(data):
-    return json.dumps({
-        "ts": data['ts'],
-        "cmd": "output",
-        "power": status_to_orders[data['output']]
-    })
